requirements.txt
Django==2.0.4
django-email-login==0.2.4
djangorestframework==3.8.2
ez-setup==0.9
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.4
